requirements.txt
requests>=2.32.0
feedparser>=6.0.11
trafilatura>=1.9.0
jinja2>=3.1.4
python-dotenv>=1.0.1
beautifulsoup4>=4.12.3
lxml>=5.2.2
lxml[html_clean]
lxml_html_clean